MedCore-HMS/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.1",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-wtf>=1.2.2",
    "psycopg2-binary>=2.9.11",
]
